refactor: move fetch into action file
